change compression method to asynchronism add decompression argument for control the method is asynchronism or synchronization use a bad method that insert receive loop for support asynchronism decompression after client call decompression syncopate mylib.package.pack3 method to pack3_compression and pack3_upload syncopate mylib.package.pack4 method to pack4_compression and pack4_upload

+++ lib/mylib/rar.py
@@ -5,12 +5,16 @@ def compression(rar, spath, files):
     for f in files:
         allpath += ' "%s"'%(spath + f)
     popen(r'del /q /f ".\temp\%s.rar"'%rar)
-    cmd = r'rar.exe a -ep -df ".\temp\%s.rar"%s'%(rar, allpath)
-    result = popen(cmd).read()
-    return '%s\n%s'%(cmd, result)
+    cmd = r'start rar.exe a -ep -df ".\temp\%s.rar"%s'%(rar, allpath)
+    popen(cmd)
+    return cmd
 
-def decompression(rar, dpath):
+def decompression(rar, dpath, async = True):
     cmd = r'rar.exe e -o+ ".\temp\%s.rar" "%s"'%(rar, dpath)
-    result = popen(cmd).read()
-    popen(r'del /q /f ".\temp\%s.rar"'%rar)
-    return '%s\n%s'%(cmd, result)
+    if async:
+        cmd = 'start %s'%cmd
+        popen(cmd)
+    else:
+        popen(cmd)
+        popen(r'del /q /f ".\temp\%s.rar"'%rar)
+    return cmd
